# api-spec.json
{"openapi":"3.0.0","paths":{"/auth/signup":{"post":{"operationId":"AuthController_signUp","summary":"","description":"signup","parameters":[],"requestBody":{"required":true,"description":"Json structure for user object","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthCrendentialsDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"properties":{"data":{"$ref":"#/components/schemas/UserPresenter"},"isArray":{"type":"boolean","default":true}}}]}}}},"401":{"description":"No authorization token was found"},"500":{"description":"Internal error"}},"tags":["Auth"]}},"/auth/signin":{"post":{"operationId":"AuthController_signIn","summary":"","description":"login","parameters":[],"requestBody":{"required":true,"description":"Json structure for user object","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthCrendentialsDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BearerTokenPresenter"}}}},"401":{"description":"No authorization token was found"},"500":{"description":"Internal error"}},"tags":["Auth"],"security":[{"bearer":[]}]}},"/trainings":{"get":{"operationId":"TrainingsController_getTrainings","parameters":[{"name":"search","required":true,"in":"query","schema":{"type":"string"}}],"requestBody":{"required":true,"description":"user","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"properties":{"data":{"$ref":"#/components/schemas/TrainingPresenter"},"isArray":{"type":"boolean","default":true}}}]}}}},"500":{"description":"Internal error"}},"tags":["Trainings"]},"post":{"operationId":"TrainingsController_createTraining","parameters":[],"requestBody":{"required":true,"description":"Json structure for training object","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrainingDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"properties":{"data":{"$ref":"#/components/schemas/TrainingPresenter"},"isArray":{"type":"boolean","default":true}}}]}}}},"500":{"description":"Internal error"}},"tags":["Trainings"]}},"/trainings/{id}":{"get":{"operationId":"TrainingsController_getTrainingById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"description":"user","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"properties":{"data":{"$ref":"#/components/schemas/TrainingPresenter"},"isArray":{"type":"boolean","default":true}}}]}}}},"500":{"description":"Internal error"}},"tags":["Trainings"]},"delete":{"operationId":"TrainingsController_deleteTraining","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"description":"user","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"properties":{"data":{"$ref":"#/components/schemas/TrainingPresenter"},"isArray":{"type":"boolean","default":true}}}]}}}},"500":{"description":"Internal error"}},"tags":["Trainings"]}},"/trainings/{id}/update":{"patch":{"operationId":"TrainingsController_updateTraining","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"description":"Json structure for training object","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrainingDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"properties":{"data":{"$ref":"#/components/schemas/TrainingPresenter"},"isArray":{"type":"boolean","default":true}}}]}}}},"500":{"description":"Internal error"}},"tags":["Trainings"]}}},"info":{"title":"E-learning courses API","description":"Nestjs API","version":"1.0","contact":{}},"tags":[{"name":"Your API Tag","description":""}],"servers":[{"url":"http://localhost:3000","description":"Local environment"}],"components":{"schemas":{"User":{"type":"object","properties":{}},"AuthCrendentialsDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"BearerTokenPresenter":{"type":"object","properties":{"accessToken":{"type":"string"}},"required":["accessToken"]},"TrainingPresenter":{"type":"object","properties":{"id":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"title":{"type":"string"},"description":{"type":"string"},"authorId":{"type":"string"}},"required":["id","createdAt","updatedAt","title","description","authorId"]},"TrainingDto":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"}},"required":["title","description"]}}}}